requirement.txt
Package                 Version
----------------------- -------------------
absl-py                 0.13.0
astor                   0.8.1
astunparse              1.6.3
cached-property         1.5.2
cachetools              4.2.2
certifi                 2021.5.30
charset-normalizer      2.0.4
cycler                  0.10.0
flatbuffers             1.12
future                  0.18.2
gast                    0.4.0
google-auth             1.34.0
google-auth-oauthlib    0.4.5
google-pasta            0.2.0
grpcio                  1.34.1
h5py                    3.1.0
idna                    3.2
imageio                 2.9.0
importlib-metadata      4.6.3
joblib                  1.0.1
Keras-Applications      1.0.8
keras-nightly           2.5.0.dev2021032900
Keras-Preprocessing     1.1.2
keras-vis               0.4.1
kiwisolver              1.3.1
Markdown                3.3.4
matplotlib              3.4.2
networkx                2.6.2
np-utils                0.5.12.1
numpy                   1.19.5
oauthlib                3.1.1
opencv-contrib-python   4.5.3.56
opencv-python           4.5.3.56
opt-einsum              3.3.0
pandas                  1.3.1
Pillow                  8.3.1
pip                     21.1.3
protobuf                3.17.3
pyasn1                  0.4.8
pyasn1-modules          0.2.8
pyparsing               2.4.7
python-dateutil         2.8.2
pytz                    2021.1
PyWavelets              1.1.1
PyYAML                  5.4.1
requests                2.26.0
requests-oauthlib       1.3.0
rsa                     4.7.2
scikit-image            0.18.2
scikit-learn            0.24.2
scipy                   1.4.1
setuptools              52.0.0.post20210125
six                     1.15.0
sklearn                 0.0
tensorboard             2.5.0
tensorboard-data-server 0.6.1
tensorboard-plugin-wit  1.8.0
tensorflow              2.5.0
tensorflow-estimator    2.5.0
termcolor               1.1.0
threadpoolctl           2.2.0
tifffile                2021.7.30
typing-extensions       3.7.4.3
urllib3                 1.26.6
Werkzeug                2.0.1
wheel                   0.36.2
wincertstore            0.2
wrapt                   1.12.1
zipp                    3.5.0
